Fixed weapons damage

+++ Assets/Scripts/Enemies/Enemy.cs
@@ -60,7 +60,7 @@ public class Enemy : MonoBehaviour {
         Rigidbody.useGravity = false;
     }
 
-    public void TakeDamage(int amount) {
+    public void TakeDamage(float amount) {
         hp -= amount;
         if (hp <= 0) {
             Die();
